tests/e2e/requirements.txt
robotframework>=7.0
robotframework-requests>=0.9
PyYAML>=6.0
